Modify escore-comps.xml and escore-packageis-list for redhat-upgrade-tool NetworkManager-openswan.
There is no package named redhat-upgrade-tool in everything repo of centos7.
then we will not use it as well. So delete it from our project.

The package NetworkManager-openswan do not exist as well. but there is
NetworkManager-libreswan instead. Centos and our iso have use it now.
So update our configuration.

Signed-off-by: Can Liu <can.liu@easystack.cn>
Acked-by: Chung-Ju Wu <jason.wu@easystack.cn>